test: 트랜잭션 분리에 따른 테스트 코드 수정
