+++ src/test/java/com/gangku/be/service/user/RegisterUserUnitTest.java
@@ -2,7 +2,6 @@ package com.gangku.be.service.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
 import com.gangku.be.domain.User;
@@ -13,13 +12,10 @@ import com.gangku.be.exception.CustomException;
 import com.gangku.be.exception.constant.AuthErrorCode;
 import com.gangku.be.exception.constant.UserErrorCode;
 import com.gangku.be.external.ai.AiApiClient;
-import com.gangku.be.repository.CategoryRepository;
-import com.gangku.be.repository.PreferredCategoryRepository;
-import com.gangku.be.repository.ReviewRepository;
 import com.gangku.be.repository.UserRepository;
 import com.gangku.be.service.UserService;
+import com.gangku.be.service.command.UserCommandService;
 import com.gangku.be.util.ai.AiTextFilterMapper;
-import com.gangku.be.util.object.FileUrlResolver;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
@@ -30,22 +26,17 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Tag("unit")
 @ExtendWith(MockitoExtension.class)
 public class RegisterUserUnitTest {
 
     @Mock private UserRepository userRepository;
-    @Mock private CategoryRepository categoryRepository;
-    @Mock private PreferredCategoryRepository preferredCategoryRepository;
-    @Mock private FileUrlResolver fileUrlResolver;
     @Mock private StringRedisTemplate stringRedisTemplate;
-    @Mock private PasswordEncoder passwordEncoder;
-    @Mock private ReviewRepository reviewRepository;
     @Mock private AiApiClient aiApiClient;
     @Mock private AiTextFilterMapper aiTextFilterMapper;
     @Mock private HashOperations<String, Object, Object> hashOperations;
+    @Mock private UserCommandService userCommandService;
 
     @InjectMocks private UserService userService;
 
@@ -63,29 +54,27 @@ public class RegisterUserUnitTest {
         TextFilterRequestDto textFilterRequestDto = mock(TextFilterRequestDto.class);
         TextFilterResponseDto textFilterResponseDto = mock(TextFilterResponseDto.class);
 
+        User expectedUser =
+                User.builder()
+                        .email("test@example.com")
+                        .nickname("정상닉네임")
+                        .build();
+
         when(stringRedisTemplate.opsForHash()).thenReturn(hashOperations);
         when(hashOperations.entries(sessionKey))
                 .thenReturn(Map.of("verified", "1", "email", "test@example.com"));
-
         when(userRepository.existsByEmail("test@example.com")).thenReturn(false);
         when(userRepository.existsByNickname("정상닉네임")).thenReturn(false);
-
         when(aiTextFilterMapper.fromSignUp(requestDto)).thenReturn(textFilterRequestDto);
         when(aiApiClient.filterText(textFilterRequestDto)).thenReturn(textFilterResponseDto);
         when(textFilterResponseDto.isAllowed()).thenReturn(true);
-
-        when(passwordEncoder.encode("plain-password")).thenReturn("encoded-password");
+        when(userCommandService.saveUser(requestDto, sessionId)).thenReturn(expectedUser);
 
         // when
-        User savedUser = userService.registerUser(requestDto, sessionId);
+        User result = userService.registerUser(requestDto, sessionId);
 
         // then
-        assertThat(savedUser.getEmail()).isEqualTo("test@example.com");
-        assertThat(savedUser.getPassword()).isEqualTo("encoded-password");
-        assertThat(savedUser.getNickname()).isEqualTo("정상닉네임");
-        assertThat(savedUser.getAge()).isEqualTo(24);
-        assertThat(savedUser.getGender()).isEqualTo("MALE");
-        assertThat(savedUser.getEnrollNumber()).isEqualTo(20);
+        assertThat(result).isEqualTo(expectedUser);
 
         verify(stringRedisTemplate, times(1)).opsForHash();
         verify(hashOperations, times(1)).entries(sessionKey);
@@ -93,19 +82,11 @@ public class RegisterUserUnitTest {
         verify(userRepository, times(1)).existsByNickname("정상닉네임");
         verify(aiTextFilterMapper, times(1)).fromSignUp(requestDto);
         verify(aiApiClient, times(1)).filterText(textFilterRequestDto);
-        verify(passwordEncoder, times(1)).encode("plain-password");
-        verify(userRepository, times(1)).save(any(User.class));
-        verify(stringRedisTemplate, times(1)).delete(sessionKey);
+        verify(userCommandService, times(1)).saveUser(requestDto, sessionId);
 
-        verifyNoInteractions(
-                categoryRepository, preferredCategoryRepository, fileUrlResolver, reviewRepository);
         verifyNoMoreInteractions(
-                userRepository,
-                stringRedisTemplate,
-                passwordEncoder,
-                aiTextFilterMapper,
-                aiApiClient,
-                hashOperations);
+                userRepository, aiTextFilterMapper, aiApiClient,
+                userCommandService, hashOperations, stringRedisTemplate);
     }
 
     @Test
@@ -131,15 +112,7 @@ public class RegisterUserUnitTest {
         verify(stringRedisTemplate, times(1)).opsForHash();
         verify(hashOperations, times(1)).entries(sessionKey);
 
-        verifyNoInteractions(
-                userRepository,
-                categoryRepository,
-                preferredCategoryRepository,
-                fileUrlResolver,
-                passwordEncoder,
-                reviewRepository,
-                aiApiClient,
-                aiTextFilterMapper);
+        verifyNoInteractions(userRepository, aiApiClient, aiTextFilterMapper, userCommandService);
         verifyNoMoreInteractions(stringRedisTemplate, hashOperations);
     }
 
@@ -168,19 +141,9 @@ public class RegisterUserUnitTest {
         verify(stringRedisTemplate, times(1)).opsForHash();
         verify(hashOperations, times(1)).entries(sessionKey);
         verify(userRepository, times(1)).existsByEmail("test@example.com");
-
         verify(userRepository, never()).existsByNickname(anyString());
-        verify(userRepository, never()).save(any());
-        verify(stringRedisTemplate, never()).delete(anyString());
 
-        verifyNoInteractions(
-                categoryRepository,
-                preferredCategoryRepository,
-                fileUrlResolver,
-                passwordEncoder,
-                reviewRepository,
-                aiApiClient,
-                aiTextFilterMapper);
+        verifyNoInteractions(aiApiClient, aiTextFilterMapper, userCommandService);
         verifyNoMoreInteractions(userRepository, stringRedisTemplate, hashOperations);
     }
 
@@ -212,17 +175,7 @@ public class RegisterUserUnitTest {
         verify(userRepository, times(1)).existsByEmail("test@example.com");
         verify(userRepository, times(1)).existsByNickname("중복닉네임");
 
-        verify(userRepository, never()).save(any());
-        verify(stringRedisTemplate, never()).delete(anyString());
-
-        verifyNoInteractions(
-                categoryRepository,
-                preferredCategoryRepository,
-                fileUrlResolver,
-                passwordEncoder,
-                reviewRepository,
-                aiApiClient,
-                aiTextFilterMapper);
+        verifyNoInteractions(aiApiClient, aiTextFilterMapper, userCommandService);
         verifyNoMoreInteractions(userRepository, stringRedisTemplate, hashOperations);
     }
 
@@ -245,7 +198,6 @@ public class RegisterUserUnitTest {
                 .thenReturn(Map.of("verified", "1", "email", "test@example.com"));
         when(userRepository.existsByEmail("test@example.com")).thenReturn(false);
         when(userRepository.existsByNickname("금칙어닉네임")).thenReturn(false);
-
         when(aiTextFilterMapper.fromSignUp(requestDto)).thenReturn(textFilterRequestDto);
         when(aiApiClient.filterText(textFilterRequestDto)).thenReturn(textFilterResponseDto);
         when(textFilterResponseDto.isAllowed()).thenReturn(false);
@@ -263,18 +215,8 @@ public class RegisterUserUnitTest {
         verify(aiTextFilterMapper, times(1)).fromSignUp(requestDto);
         verify(aiApiClient, times(1)).filterText(textFilterRequestDto);
 
-        verify(passwordEncoder, never()).encode(anyString());
-        verify(userRepository, never()).save(any());
-        verify(stringRedisTemplate, never()).delete(anyString());
-
-        verifyNoInteractions(
-                categoryRepository, preferredCategoryRepository, fileUrlResolver, reviewRepository);
+        verifyNoInteractions(userCommandService);
         verifyNoMoreInteractions(
-                userRepository,
-                stringRedisTemplate,
-                aiTextFilterMapper,
-                aiApiClient,
-                passwordEncoder,
-                hashOperations);
+                userRepository, stringRedisTemplate, aiTextFilterMapper, aiApiClient, hashOperations);
     }
 }
